スナップ処理を別ファイルに書き出し
