+++ src/components/OrigamiPost/hooks/useSelectPoints/index.tsx
@@ -8,8 +8,8 @@ import { renderPoint } from "../../logics/renderPoint";
 import { currentStepAtom } from "../../atoms/currentStepAtom";
 import { inputStepObjectAtom } from "../../atoms/inputStepObjectAtom";
 import { useAtom, useAtomValue } from "jotai";
-import { renderHighlightPoint } from "../../logics/renderPoint";
 import { useInitialRender } from "./useInitialRender";
+import { useMouseMove } from "./useMouseMove";
 
 type UseSelectPoints = (props: {
   canvasRef: React.MutableRefObject<HTMLCanvasElement | null>;
@@ -40,7 +40,6 @@ export const useSelectPoints: UseSelectPoints = ({
   // スナップ用の状態を追加
   const [highlightedVertex, setHighlightedVertex] =
     React.useState<THREE.Vector3 | null>(null);
-  const SNAP_THRESHOLD = 0.1; // スナップする距離の閾値
 
   // 初期描画のロジックを分離
   useInitialRender({
@@ -53,123 +52,17 @@ export const useSelectPoints: UseSelectPoints = ({
     origamiColor,
   });
 
-  // マウスの移動を監視する処理を追加
-  useEffect(() => {
-    if (inputStep !== "axis") return;
-    const canvas = canvasRef.current!;
-    const scene = sceneRef.current!;
-    const camera = cameraRef.current!;
-    const renderer = rendererRef.current!;
-    const raycaster = raycasterRef.current!;
-
-    const mouseMoveListener = (event: MouseEvent) => {
-      const sizes = {
-        width: window.innerWidth - 320,
-        height: window.innerHeight,
-      };
-
-      // マウス座標の正規化
-      const mouse = new THREE.Vector2();
-      mouse.x = (event.clientX / sizes.width) * 2 - 1;
-      mouse.y = -(event.clientY / sizes.height) * 2 + 1;
-
-      // 全ての板の頂点と辺の中心点を取得
-      let closestVertex: THREE.Vector3 | null = null;
-      let minDistance = Infinity;
-
-      // 頂点と辺の中心点のチェック
-      fixBoards.forEach((board) => {
-        // 頂点のチェック
-        board.forEach((vertex) => {
-          const vertexVector = new THREE.Vector3(
-            vertex[0],
-            vertex[1],
-            vertex[2]
-          );
-          // 頂点をスクリーン座標(マウスの位置)に変換
-          const vertexScreenPos = vertexVector.clone().project(camera);
-          // マウスとの距離を計算
-          const distance = new THREE.Vector2(mouse.x, mouse.y).distanceTo(
-            new THREE.Vector2(vertexScreenPos.x, vertexScreenPos.y)
-          );
-          // 距離が閾値より小さい場合は、頂点を更新
-          if (distance < SNAP_THRESHOLD && distance < minDistance) {
-            minDistance = distance;
-            closestVertex = vertexVector;
-          }
-        });
-
-        // 辺の中心点のチェック
-        for (let i = 0; i < board.length; i++) {
-          const nextIndex = (i + 1) % board.length;
-          const currentVertex = board[i];
-          const nextVertex = board[nextIndex];
-
-          // 辺の中心点を計算
-          const centerPoint = new THREE.Vector3(
-            (currentVertex[0] + nextVertex[0]) / 2,
-            (currentVertex[1] + nextVertex[1]) / 2,
-            (currentVertex[2] + nextVertex[2]) / 2
-          );
-
-          const centerScreenPos = centerPoint.clone().project(camera);
-          const distance = new THREE.Vector2(mouse.x, mouse.y).distanceTo(
-            new THREE.Vector2(centerScreenPos.x, centerScreenPos.y)
-          );
-
-          if (distance < SNAP_THRESHOLD && distance < minDistance) {
-            minDistance = distance;
-            closestVertex = centerPoint;
-          }
-        }
-      });
-
-      // 頂点も辺の中心点も見つからない場合は、エッジ上の最近点を探す
-      if (!closestVertex) {
-        // Raycasterを使用してエッジとの交差を確認
-        raycaster.setFromCamera(mouse, camera);
-        const edges = scene.children.filter(
-          (child) => child.type === "LineSegments"
-        );
-        const intersects = raycaster.intersectObjects(edges, true);
-
-        if (intersects.length > 0) {
-          closestVertex = intersects[0].point;
-        }
-      }
-
-      // 既存のハイライトを削除
-      scene.children = scene.children.filter(
-        (child) => child.name !== "HighlightPoint"
-      );
-
-      // 新しいハイライトを描画
-      if (closestVertex) {
-        renderHighlightPoint({
-          scene,
-          point: [closestVertex.x, closestVertex.y, closestVertex.z],
-        });
-        setHighlightedVertex(closestVertex);
-      } else {
-        setHighlightedVertex(null);
-      }
-
-      renderer.render(scene, camera);
-    };
-
-    canvas.addEventListener("mousemove", mouseMoveListener);
-    return () => {
-      canvas.removeEventListener("mousemove", mouseMoveListener);
-    };
-  }, [
+  // マウス移動時の処理を分離
+  useMouseMove({
     inputStep,
     canvasRef,
     sceneRef,
     cameraRef,
     rendererRef,
-    fixBoards,
     raycasterRef,
-  ]);
+    fixBoards,
+    setHighlightedVertex,
+  });
 
   // モデル上の点を選択できるようにする処理
   useEffect(() => {
